Removed test_old.rs, moved its contents into appropriate files
