+++ src/main.rs
@@ -6,9 +6,7 @@ pub mod simulation;
 pub mod util;
 pub mod juice_renderer;
 pub mod error;
-// pub mod test and ui;
-
-pub mod test_old;
+pub mod test;
 pub mod ui;
 
 use simulation::Simulation;
