+++ src/test/mod.rs
@@ -1,0 +1,4 @@
+pub mod test_physics;
+pub mod test_renderer;
+pub mod test_state_manager;
+pub mod test_ui;
